Put string in DOM instead of object
JSON stringification is non-deterministic. Since the server and browser
can create different strings, we were setting ourselves up for checksum
errors (and got them in IE10). Now we put the JSON in the DOM as a
string and use that string exactly when rendering on the browser side.
This guarantees they’re identical.